Refactor to use vite + ts
