--- inventree_bulk_plugin/templates/panels/category-detail/create-bulk.js
@@ -1,10 +1,0 @@
-{% include "components/generateBulkCreateView.js" %}
-
-generateBulkCreateView({
-  target: document.getElementById("bulk-create-category-preact-root"),
-  createURL: "{% url 'plugin:inventree-bulk-plugin:bulkcreatecategory' pk=category.id %}",
-  name: "categories",
-  defaultSchema: null,
-  tableSelector: "#bulk-create-category-preview-table",
-  templateType: "PART_CATEGORY",
-});
