--- inventree_bulk_plugin/templates/panels/stock-location-detail/create-bulk.html
@@ -1,13 +1,0 @@
-{% load static %}
-<script src="{% static 'inventree-bulk-plugin/preact.js' %}"></script>
-<script src="{% static 'inventree-bulk-plugin/htm.js' %}"></script>
-
-<div id="bulk-create-location-preact-root"></div>
-
-<table id="bulk-create-location-preview-table" class="mt-3"></table>
-
-<style>
-  .loadingindicator {
-    display: none;
-  }
-</style>
