--- inventree_bulk_plugin/templates/panels/stock-location-detail/create-bulk.js
@@ -1,10 +1,0 @@
-{% include "components/generateBulkCreateView.js" %}
-
-generateBulkCreateView({
-  target: document.getElementById("bulk-create-location-preact-root"),
-  createURL: "{% url 'plugin:inventree-bulk-plugin:bulkcreatelocation' pk=location.id %}",
-  name: "locations",
-  defaultSchema: null,
-  tableSelector: "#bulk-create-location-preview-table",
-  templateType: "STOCK_LOCATION",
-});
